core/wsgi.py
import os
from dotenv import load_dotenv

# Cargar el .env desde la raíz del proyecto o donde esté ubicado
load_dotenv(dotenv_path=os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env'))

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'core.settings')

from django.core.wsgi import get_wsgi_application

application = get_wsgi_application()
