requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
uritemplate==4.2.0
